Conflicts:
	CMakeLists.txt

+++ src/engine/demo.h
@@ -10,7 +10,7 @@ enum
 	MAX_TIMELINE_MARKERS=64
 };
 
-const double g_aSpeeds[] = {0.1, 0.25, 0.5, 0.75, 1.0, 1.5, 2.0, 4.0, 8.0};
+const float g_aSpeeds[] = {0.1, 0.25, 0.5, 0.75, 1.0, 1.5, 2.0, 4.0, 8.0};
 
 struct CDemoHeader
 {
